refactor: 변수, 메서드명 리팩토링 및 로직 변경중...
