+++ src/main/java/com/capstone/cschatbot/chat/controller/ChatController.java
@@ -1,7 +1,7 @@
 package com.capstone.cschatbot.chat.controller;
 
 import com.capstone.cschatbot.chat.dto.ChatDto;
-import com.capstone.cschatbot.chat.service.ChatService;
+import com.capstone.cschatbot.chat.service.chat.ChatService;
 import com.capstone.cschatbot.common.dto.ApiResponse;
 import com.capstone.cschatbot.common.enums.CustomResponseStatus;
 import com.capstone.cschatbot.config.security.service.PrincipalDetails;
@@ -19,44 +19,45 @@ public class ChatController {
     private final ChatService chatService;
 
     @GetMapping("/initial/chat")
-    public ResponseEntity<ApiResponse<ChatDto.Response.Chat>> initialChat(
+    public ResponseEntity<ApiResponse<ChatDto.Response.Chat>> initiateCSChat(
             @AuthenticationPrincipal PrincipalDetails principalDetails,
             @RequestParam("topic") String topic) {
 
         return ResponseEntity.ok().body(ApiResponse.createSuccess(
-                chatService.csInitChat(principalDetails.getMemberId(), topic),
+                chatService.initiateCSChat(principalDetails.getMemberId(), topic),
                 CustomResponseStatus.SUCCESS)
         );
     }
 
     @GetMapping("/self/initial/chat")
-    public ResponseEntity<ApiResponse<ChatDto.Response.Chat>> selfInitialChat(
+    public ResponseEntity<ApiResponse<ChatDto.Response.Chat>> initiateSelfIntroChat(
             @AuthenticationPrincipal PrincipalDetails principalDetails,
-            @ModelAttribute ChatDto.Request.SelfChat chat
+            @ModelAttribute ChatDto.Request.SelfIntroChat chat
     ) {
-        log.info("question : {} content : {}", chat.getQuestion(), chat.getContent());
+
         return ResponseEntity.ok().body(ApiResponse.createSuccess(
-                chatService.selfInitChat(principalDetails.getMemberId(), chat),
+                chatService.initiateSelfIntroChat(principalDetails.getMemberId(), chat),
                 CustomResponseStatus.SUCCESS)
         );
     }
 
     @GetMapping("/chat")
-    public ResponseEntity<ApiResponse<ChatDto.Response.EvaluationAndQuestion>> chat(
+    public ResponseEntity<ApiResponse<ChatDto.Response.EvaluationAndQuestion>> processChat(
             @AuthenticationPrincipal PrincipalDetails principalDetails,
             @RequestParam ChatDto.Request.Chat prompt
     ) {
         return ResponseEntity.ok().body(ApiResponse.createSuccess(
-                chatService.chat(principalDetails.getMemberId(), prompt),
+                chatService.processChat(principalDetails.getMemberId(), prompt),
                 CustomResponseStatus.SUCCESS)
         );
     }
 
+    // TODO : 채팅 종료시 답변 평가를 보여준다면 로직이 많이 변할 예정
     @PostMapping("/end/chat")
-    public ResponseEntity<ApiResponse<String>> endChat(
+    public ResponseEntity<ApiResponse<String>> terminateChat(
             @AuthenticationPrincipal PrincipalDetails principalDetails
     ) {
-        chatService.endChat(principalDetails.getMemberId());
+        chatService.terminateChat(principalDetails.getMemberId());
         return ResponseEntity.ok().body(ApiResponse.createSuccess("채팅이 종료되었습니다.", CustomResponseStatus.SUCCESS));
     }
