+++ src/main/java/com/capstone/cschatbot/chat/entity/enums/GPTRoleType.java
@@ -1,0 +1,16 @@
+package com.capstone.cschatbot.chat.entity.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum GPTRoleType {
+    USER("user"),
+    ASSISTANT("assistant"),
+    SYSTEM("system");
+
+    private String role;
+
+    GPTRoleType(String role) {
+        this.role = role;
+    }
+}
